pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yokin-chosa"
version = "0.1.0"
description = "相続税申告用 預金調査ツール"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "python-multipart>=0.0.6",
    "jinja2>=3.1",
    "pandas>=2.0",
    "openpyxl>=3.1",
    "pdfplumber>=0.10",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
desktop = [
    "pywebview>=4.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "httpx>=0.25",
]

[project.scripts]
yokin-chosa = "yokin_chosa.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
